Created the workflow tangelo plugin
